* SphericalBesselGenerator optimizations
--HG--
extra : convert_revision : svn%3A15b6bd1d-d91c-0410-8971-a6cc2a4f8591/epdp/trunk%401127

+++ Defs.hpp
@@ -7,8 +7,8 @@
 #include <cmath>
 
 typedef double Real;
-typedef int Integer;
-typedef unsigned int UnsignedInteger;
+typedef long int Integer;
+typedef unsigned long int UnsignedInteger;
 typedef size_t Index;
